+++ src/ui/keymaps.hpp
@@ -71,8 +71,8 @@ Keybinds parseKeybinds()
    */
   auto reportError = [](const std::string& field, const std::string_view& key)
   {
-    std::cerr << "Error: Unsupported or empty keybind '" << key << "' detected for field '" << field
-              << "'. Please check your configuration." << std::endl;
+    std::cerr << "** Error: Unsupported or empty keybind '" << key << "' detected for field '" << field
+              << "'. Modify or remove that keybind field. **" << std::endl;
     std::exit(EXIT_FAILURE); /**< Exit the program on invalid keybind */
   };